requirements.txt
# Core Django and Database
Django==4.2.7
# Using SQLite for development (default with Django)
# psycopg2-binary==2.9.7  # Uncomment when PostgreSQL is set up

# Environment and Configuration
python-decouple==3.8

# Media and Image Processing
Pillow==10.0.0

# Authentication and Security
django-phonenumber-field==7.1.0
phonenumbers==8.13.19

# Development Tools
django-debug-toolbar==4.2.0

# Internationalization and Localization
django-modeltranslation==0.18.11

# API and AJAX Support
djangorestframework==3.14.0

# Caching (will be used later for optimization)
redis==4.6.0
django-redis==5.3.0

# Form Enhancements
django-widget-tweaks==1.4.12

# Time and Date
pytz==2023.3

# Development and Testing
factory-boy==3.3.0
coverage==7.3.0